+++ src/backend/core/api/viewsets.py
@@ -3085,6 +3085,7 @@ class ConfigView(drf.views.APIView):
             "CONVERSION_FILE_EXTENSIONS_ALLOWED",
             "CONVERSION_FILE_MAX_SIZE",
             "CONVERSION_UPLOAD_ENABLED",
+            "DOCUMENT_IMAGE_MAX_SIZE",
             "ENVIRONMENT",
             "FRONTEND_CSS_URL",
             "FRONTEND_HOMEPAGE_FEATURE_ENABLED",